refactoring and conditioning to avoid know errors
